# Conflicts:
#	TrainAndTest/fbank/LSTMs/__init__.py
#	TrainAndTest/fbank/LSTMs/train_alstm_kaldi.py
#	TrainAndTest/fbank/LSTMs/train_alstm_tuple.py
#	TrainAndTest/fbank/LSTMs/train_lstm_kaldi.py
#	TrainAndTest/fbank/LSTMs/train_lstm_tuple.py
#	TrainAndTest/fbank/ResNets/__init__.py
#	TrainAndTest/fbank/ResNets/aidata_resnet.yaml
#	TrainAndTest/fbank/ResNets/aidata_resnet_mixup.yaml
#	TrainAndTest/fbank/ResNets/cnc1_resnet.yaml
#	TrainAndTest/fbank/ResNets/cnc1_resnet_mixup.yaml
#	TrainAndTest/fbank/ResNets/cnc1_vox1.yaml
#	TrainAndTest/fbank/ResNets/fine_exres_cnc.py
#	TrainAndTest/fbank/ResNets/train_amsoftmax_res10.py
#	TrainAndTest/fbank/ResNets/train_amsoftmax_res34.py
#	TrainAndTest/fbank/ResNets/train_asires34_aug.py
#	TrainAndTest/fbank/ResNets/train_asoftmax_res34.py
#	TrainAndTest/fbank/ResNets/train_exres_kaldi.py
#	TrainAndTest/fbank/ResNets/train_exres_var.py
#	TrainAndTest/fbank/ResNets/train_res_kaldi.py
#	TrainAndTest/fbank/ResNets/train_resnet10.py
#	TrainAndTest/fbank/ResNets/vox1_cnc1.yaml
#	TrainAndTest/fbank/ResNets/vox1_resnet.yaml
#	TrainAndTest/fbank/ResNets/vox1_resnet_mixup.yaml
#	TrainAndTest/fbank/ResNets/vox2_resnet.yaml
#	TrainAndTest/fbank/ResNets/vox2_resnet_mixup.yaml
#	TrainAndTest/fbank/SuResCNN/train_asoftmax_sures10.py
#	TrainAndTest/fbank/SuResCNN/train_rescnn10.py
#	TrainAndTest/fbank/SuResCNN/train_surescnn10.py
#	TrainAndTest/fbank/TDNNs/__init__.py
#	TrainAndTest/fbank/TDNNs/cnc1_tdnn.yaml
#	TrainAndTest/fbank/TDNNs/train_astdnn_kaldi.py
#	TrainAndTest/fbank/TDNNs/train_etdnn_kaldi.py
#	TrainAndTest/fbank/TDNNs/train_sitdnn.py
#	TrainAndTest/fbank/TDNNs/train_tdnn_kaldi.py
#	TrainAndTest/fbank/TDNNs/train_tdnn_var.py
#	TrainAndTest/fbank/TDNNs/train_tdnn_vox1.py
#	TrainAndTest/fbank/TDNNs/vox1_tdnn.yaml
#	TrainAndTest/fbank/TDNNs/vox1_tdnn_drop.yaml
#	TrainAndTest/fbank/TDNNs/vox2_ecapa.yaml
#	TrainAndTest/fbank/TDNNs/vox2_ecapa_mixup.yaml
#	TrainAndTest/fbank/TDNNs/vox2_tdnn.yaml
#	TrainAndTest/fbank/TDNNs/vox2_tdnn_mixup.yaml
#	TrainAndTest/fbank/train_dom.sh
#	TrainAndTest/fbank/train_resnet.sh
#	TrainAndTest/fbank/train_tdnn.sh
#	TrainAndTest/spectrogram/Contest.py
#	TrainAndTest/spectrogram/ResNets/vox1_resnet10/baseline.yaml
#	TrainAndTest/spectrogram/ResNets/vox1_resnet18/drop_v2_rgn.yaml
#	TrainAndTest/spectrogram/ResNets/vox1_resnet34.yaml
#	TrainAndTest/spectrogram/ResNets/vox2_resnet.yaml
#	TrainAndTest/spectrogram/train_cnn.sh
#	TrainAndTest/spectrogram/train_dom.sh
#	TrainAndTest/spectrogram/train_domres_kaldi.py
#	TrainAndTest/spectrogram/train_egs_var.py
#	TrainAndTest/spectrogram/train_kd.sh
#	TrainAndTest/spectrogram/train_lmdb.py
#	TrainAndTest/spectrogram/train_lores.sh
#	TrainAndTest/spectrogram/train_lores10_kaldi.py
#	TrainAndTest/spectrogram/train_lores10_var.py
#	TrainAndTest/spectrogram/train_lores_timit.sh
#	TrainAndTest/spectrogram/train_multi.sh
#	TrainAndTest/spectrogram/train_resnet.py
#	TrainAndTest/spectrogram/train_resnet20.sh
#	TrainAndTest/spectrogram/train_resnet20_kaldi.py
#	TrainAndTest/wav/mani_layers/vox1_resnet_mixup_0.yaml
#	TrainAndTest/wav/mani_layers/vox1_resnet_mixup_01.yaml
#	TrainAndTest/wav/mani_layers/vox1_resnet_mixup_012.yaml
#	TrainAndTest/wav/mani_layers/vox1_resnet_mixup_02.yaml
#	TrainAndTest/wav/mani_layers/vox1_resnet_mixup_023.yaml
#	TrainAndTest/wav/mani_layers/vox1_resnet_mixup_1.yaml
#	TrainAndTest/wav/mani_layers/vox1_resnet_mixup_123.yaml
#	TrainAndTest/wav/mani_layers/vox1_resnet_mixup_1234.yaml
#	TrainAndTest/wav/mani_layers/vox1_resnet_mixup_2.yaml
#	TrainAndTest/wav/mani_layers/vox1_resnet_mixup_234.yaml
#	TrainAndTest/wav/mani_layers/vox1_resnet_mixup_345.yaml
#	TrainAndTest/wav/mani_layers/vox1_resnet_mixup_456.yaml
#	TrainAndTest/wav/mani_layers/vox1_resnet_mixup_567.yaml
#	TrainAndTest/wav/mani_layers/vox1_resnet_mixup_67.yaml
#	TrainAndTest/wav/mani_layers/vox1_resnet_mixup_7.yaml
#	TrainAndTest/wav/train_filter.sh
#	TrainAndTest/wav/train_tdnn.sh
#	TrainAndTest/wav/vox1_resnet.yaml
#	TrainAndTest/wav/vox1_resnet_mixup.yaml
#	TrainAndTest/wav/vox2_ecapa.yaml

+++ Process_Data/audio_processing.py
@@ -439,7 +439,8 @@ def read_Waveform(filename):
     """
     # audio, sr = librosa.load(filename, sr=sample_rate, mono=True)
     # audio = audio.flatten()
-    audio, sample_rate = sf.read(filename, dtype='int16')
+    # audio, sample_rate = sf.read(filename, dtype='int16')
+    audio, sample_rate = sf.read(filename, dtype='float32')
 
     return audio.astype(np.float32).reshape(1, -1)
